Revert "[Dynamo][12/N] Remove allowed_functions.py (#116401)"
This reverts commit 7e12e722af5e4d085c9f7ced90448d5745eb4c6e.
