+++ torch/_dynamo/decorators.py
@@ -2,7 +2,7 @@ from typing import TYPE_CHECKING
 
 import torch
 from torch.utils._python_dispatch import is_traceable_wrapper_subclass
-from . import trace_rules, variables
+from . import allowed_functions, trace_rules, variables
 from .eval_frame import DisableContext, innermost_fn, RunOnlyContext
 from .exc import IncorrectUsage
 
@@ -92,8 +92,8 @@ def allow_in_graph(fn):
         return [allow_in_graph(x) for x in fn]
     assert callable(fn), "allow_in_graph expects a callable"
     if trace_rules.lookup(fn) != variables.TorchInGraphFunctionVariable:
-        trace_rules._disallowed_callable_ids.remove(id(fn))
-        trace_rules._allowed_callable_ids.add(id(fn))
+        allowed_functions._disallowed_callable_ids.remove(id(fn))
+        allowed_functions._allowed_callable_ids.add(id(fn))
     return fn
 
 
@@ -105,14 +105,14 @@ def _disallow_in_graph_helper(throw_if_not_allowed):
         if (
             throw_if_not_allowed
             and trace_rules.lookup(fn) != variables.TorchInGraphFunctionVariable
-            and fn not in trace_rules._allowed_callable_ids
+            and fn not in allowed_functions._allowed_callable_ids
         ):
             raise IncorrectUsage(
                 "disallow_in_graph is expected to be used on an already allowed callable (like torch.* ops). "
                 "Allowed callables means callables that TorchDynamo puts as-is in the extracted graph."
             )
-        trace_rules._allowed_callable_ids.remove(id(fn))
-        trace_rules._disallowed_callable_ids.add(id(fn))
+        allowed_functions._allowed_callable_ids.remove(id(fn))
+        allowed_functions._disallowed_callable_ids.add(id(fn))
         return fn
 
     return inner
@@ -313,4 +313,4 @@ def _allow_in_graph_einops():
             allow_in_graph(einops.unpack)  # available since einops 0.6.0
 
 
-trace_rules.add_module_init_func("einops", _allow_in_graph_einops)
+allowed_functions.add_module_init_func("einops", _allow_in_graph_einops)
