+++ torch/_dynamo/variables/misc.py
@@ -450,7 +450,7 @@ class AutogradFunctionVariable(VariableTracker):
         args: "List[VariableTracker]",
         kwargs: "Dict[str, VariableTracker]",
     ):
-        from ..trace_rules import is_callable_allowed
+        from ..allowed_functions import is_callable_allowed
         from .builder import wrap_fx_proxy
 
         if name == "apply":
